drivers: mtk: separate sinegen control from memif
Sinegen control part is platform dependent and optional for debug
purporse. On the other hand, memif control driver which is platform
independent could be reused. As a result, sinegen control driver is
separated from afe-memif.c.
