+++ src/drivers/mediatek/mt8186/CMakeLists.txt
@@ -2,3 +2,7 @@
 
 add_local_sources(sof afe-memif.c afe-dai.c afe-drv.c)
 
+if(CONFIG_TEST_SGEN)
+add_local_sources(sof afe-sgen.c)
+endif()
+
