+++ src/include/sof/drivers/afe-sgen.h
@@ -1,0 +1,15 @@
+/* SPDX-License-Identifier: BSD-3-Clause */
+/*
+ * Copyright(c) 2023 MediaTek. All rights reserved.
+ *
+ * Author: Trevor Wu <trevor.wu@mediatek.com>
+ */
+
+#ifndef __SOF_DRIVERS_AFE_SGEN_H__
+#define __SOF_DRIVERS_AFE_SGEN_H__
+
+void afe_sinegen_enable(void);
+void afe_sinegen_disable(void);
+
+#endif /* __SOF_DRIVERS_AFE_SGEN_H__ */
+
